+++ lib/src/action_callback.dart
@@ -1,0 +1,1 @@
+typedef ActionCallback<T> = void Function(T key);